ml/requirements.txt
# Core ML Libraries
xgboost==2.0.3
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# Feature Engineering
scipy==1.12.0

# Model Optimization
optuna==3.5.0
shap==0.44.1
boruta==0.3

# API & Serving
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0

# Data Sources
python-binance==1.0.19
redis==5.0.1
requests==2.31.0

# MLOps
mlflow==2.9.2
joblib==1.3.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Database
sqlalchemy==2.0.25

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Deep Learning (Phase 2)
torch==2.5.1
torchvision==0.20.1
pillow==10.2.0

# Image Processing (Phase 2)
scikit-image==0.22.0
opencv-python==4.10.0.84

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
